# Glob: fix performance