@coagmano I think this PR addresses the issues you raised in https://github.com/coagmano/meteor-stylus/pull/6#issuecomment-446862256. I’m making sure the plugins are never implicitly loaded (so one would always need to `@import 'rupture'` etc.) and I’m filtering out plugin paths from the lookups. I think this should eliminate all unnecessary path lookups, e.g. for `<plugin path>/<imported file>`. Do you mind trying this out in your bigger app and letting me know how it goes? And if it fails, can you update the example repo accordingly, and give me write access to it?

I can’t figure out how to write a test case for the “local package that exports Stylus” use case as described by @hwillson in https://github.com/meteor/meteor/pull/9272#issuecomment-348249629. Is there a way to have a test local package that can be included inside this package and loaded into Meteor’s testing app during package testing?