@coagmano 2019-01-08
I've tested with and without local packages that add `.styl` files and they both work, and build times are definitely down from #6, but still higher overall (all times are after fresh `meteor reset`):

### Without local packages:

| Run  | Build time |
| ------------- | ------------- |
| Before Glob  | 2,436ms  |
| After Glob  | 59,310ms  |
| After #7 (without `@import 'nib')  | 642ms  |
| After #7 (with `@import 'nib')  | 4,029ms  |

### With (many) local packages:

| Run  | Build time |
| ------------- | ------------- |
| Before Glob  | 19,731ms  |
| After Glob  | Gave up after 10+ minutes  |
| After #7 (without `@import 'nib')  | 101,092ms  |
| After #7 (with `@import 'nib')  | Gave up after 10+ minutes |

I have a feeling that there's improvements to be made, since the performance penalty is still large, and even worse when importing nib.

It's probably good enough to ship as a major version with a note in the readme and fix the perf issues later. 
I'll just keep using the current version on my projects till perf gets fixed