@GeoffreyBooth 2019-01-08
So after #7, performance is _faster_ with no local packages, and twice as slow (or a 2 sec penalty, not sure if it will scale linearly) with `@import 'nib'`. Am I reading this correctly?

I think that Nib isn’t a plugin that needs to be imported, it should always be included. Can you confirm that you’re using something that Nib provides, and it works without the explicit import?

And if you can post a project that replicates the issue, I’m happy to look into this more (but I’m not sure when).