@GeoffreyBooth 2019-01-08
Also does anyone know how to take a string like `{coagmano:stylus}` and get the path to that package, from within a Meteor build compiler plugin? @benjamn?