@coagmano 2019-01-08
I've done a bit of work with [import-js](https://github.com/Galooshi/import-js)'s meteor support, which can parse similar strings in js files, so I can port some of it's logic over. It's just a matter of time, and I don't have a need for the feature yet!